pyproject.toml
[project]
name = "agentcore"
version = "0.1.0"
description = "AgentCore Strands Agent with MCP integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=0.7.0",
    "bedrock-agentcore>=0.1.0",
    "mcp>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["my_strands_agent"]

[tool.uv.workspace]
members = [
    "gateway",
]
